Training yields RuntimeError: Double can't be cast to desired Long. Fix?

+++ training.py
@@ -33,10 +33,10 @@ def train_fn(loader, model, optimizer, loss_fn, scaler):
     for batch_idx, (data, targets) in enumerate(loop):
         # print('enum(loop)')
         # data = data.to(device=DEVICE)
-        data = data.double().unsqueeze(1).to(device=DEVICE)
+        data = data.float().unsqueeze(1).to(device=DEVICE)
         # print('data success')
         # print(f'data.shape = {data.shape}')
-        targets = targets.double().unsqueeze(1).to(device=DEVICE)
+        targets = targets.float().unsqueeze(1).to(device=DEVICE)
         # print('targets success')
         # First consider the forward training path. This means calculate the
         # the predictions and determine the resultung error using the loss_fn.
@@ -51,7 +51,7 @@ def train_fn(loader, model, optimizer, loss_fn, scaler):
             # print('autocast success')
             predictions = model(data)
             # print('model(data) success')
-            loss = loss_fn(predictions, targets.double())
+            loss = loss_fn(predictions, targets.float())
             # print(f'The current mean average loss is: {loss}.')
 
         # Next consider the backward training path, especially the corresponding
